+++ bacnet-stack/demo/object/av.c
@@ -159,7 +159,7 @@ int Analog_Value_Encode_Property_APDU(uint8_t * apdu,
     int apdu_len = 0;           /* return value */
     BACNET_BIT_STRING bit_string;
     BACNET_CHARACTER_STRING char_string;
-    float real_value = 1.414;
+    float real_value = (float)1.414;
     unsigned object_index = 0;
     unsigned i = 0;
     bool state = false;
